+++ src/live_view.cpp
@@ -128,7 +128,7 @@ void live_view::print_items(std::vector<item> &items, int &line) const
     }
 
     int last_line = height - START_LINE - 1;
-    bool will_overflow = line-1 + (ssize_t)item_names.size() > last_line;
+    bool will_overflow = line - 1 + (int)item_names.size() > last_line;
 
     for (std::map<std::string, int>::iterator it = item_names.begin();
          it != item_names.end() && (!will_overflow || line < last_line); it++) {
